+++ src/sieve.rs
@@ -212,7 +212,11 @@ pub const fn sieve<const N: usize>() -> [bool; N] {
 
 /// The error returned by [`sieve_lt`] and [`sieve_geq`] if the input
 /// is invalid or does not work to sieve the requested range.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+///
+/// Only implements the [`Error`](std::error::Error) trait
+/// if the `std` feature is enabled.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub enum SieveError {
     /// The limit was less than or equal to `N` (for `sieve_lt`).
     TooSmallLimit,